documentation improved password checking

+++ src/main/java/sample/ciphering/cipherManagers/CipherManager.java
@@ -7,5 +7,5 @@ public interface CipherManager {
     CipherJobExecutor getJobExecutor();
     String getSourcePath();
     String getDestinationPath();
-    int getFileSize();
+    float getFileSize();
 }
